refactor(web): move marketplace contract to contracts package
