--- web/app/components/plugins/marketplace/types.ts
@@ -1,40 +1,0 @@
-export type SearchParamsFromCollection = {
-  query?: string
-  sort_by?: string
-  sort_order?: string
-}
-
-export type MarketplaceCollection = {
-  name: string
-  label: Record<string, string>
-  description: Record<string, string>
-  rule: string
-  created_at: string
-  updated_at: string
-  searchable?: boolean
-  search_params?: SearchParamsFromCollection
-}
-
-export type PluginsSearchParams = {
-  query: string
-  page?: number
-  page_size?: number
-  sort_by?: string
-  sort_order?: string
-  category?: string
-  tags?: string[]
-  exclude?: string[]
-  type?: 'plugin' | 'bundle'
-}
-
-export type PluginsSort = {
-  sortBy: string
-  sortOrder: string
-}
-
-export type CollectionsAndPluginsSearchParams = {
-  category?: string
-  condition?: string
-  exclude?: string[]
-  type?: 'plugin' | 'bundle'
-}
