+++ web/app/components/plugins/plugin-page/plugin-tasks/components/error-plugin-item.tsx
@@ -38,7 +38,7 @@ const ErrorPluginItem: FC<ErrorPluginItemProps> = ({ plugin, getIconUrl, languag
       const manifest: Plugin = {
         plugin_id: plugin.plugin_id,
         type: info.category as Plugin['type'],
-        category: info.category,
+        category: info.category as Plugin['category'],
         name: name!,
         org: org!,
         version: info.latest_version,
