+++ web/app/components/plugins/types.ts
@@ -224,12 +224,6 @@ export type PluginDetail = {
   alternative_plugin_id: string
 }
 
-export type PluginInfoFromMarketPlace = {
-  category: PluginCategoryEnum
-  latest_package_identifier: string
-  latest_version: string
-}
-
 export type Plugin = {
   type: 'plugin' | 'bundle' | 'model' | 'extension' | 'tool' | 'agent_strategy' | 'datasource' | 'trigger'
   org: string
@@ -452,22 +446,6 @@ export type InstalledPluginCategoryListResponse = {
 
 export type UninstallPluginResponse = {
   success: boolean
-}
-
-export type PluginsFromMarketplaceResponse = {
-  plugins: Plugin[]
-  bundles?: Plugin[]
-  total: number
-}
-export type PluginsFromMarketplaceByInfoResponse = {
-  list: {
-    plugin: Plugin
-    version: {
-      plugin_name: string
-      plugin_org: string
-      unique_identifier: string
-    }
-  }[]
 }
 
 export type GitHubItemAndMarketPlaceDependency = {
