--- web/contract/base.ts
@@ -1,3 +1,0 @@
-import { oc } from '@orpc/contract'
-
-export const base = oc.$route({ inputStructure: 'detailed' })
