--- web/types/marketplace-template.ts
@@ -1,11 +1,0 @@
-export type MarketplaceTemplate = {
-  id: string
-  template_name: string
-  overview: string
-  icon: string
-  icon_background: string
-  icon_file_key: string
-  publisher_unique_handle: string
-  usage_count: number
-  categories: string[]
-}
